@@ -75,15 +75,19 @@
     "passport-local": "^1.0.0",
     "path": "^0.12.7",
     "pg": "^7.0.2",
+    "prop-types": "^15.6.0",
     "ramda": "^0.24.1",
-    "react": "^15.4.1",
+    "react": "^16.0.0",
     "react-datetime": "^2.9.0",
-    "react-dom": "^15.4.1",
+    "react-dom": "^16.0.0",
+    "react-dom-factories": "^1.0.0",
     "react-elm-components": "^1.0.1",
     "react-hammerjs": "^0.5.0",
-    "react-redux": "^5.0.1",
-    "react-router-dom": "4.0.0-beta.3",
+    "react-redux": "^5.0.6",
+    "react-router-dom": "^4.2.2",
     "react-router-redux": "^5.0.0-alpha.4",
+    "react-selectize": "^3.0.1",
+    "react-transition-group": "^1.1.2",
     "reduce-reducers": "^0.1.2",
     "redux": "^3.5.2",
     "redux-thunk": "^2.1.0",
